DESD_BRFN/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
boto3==1.42.67
botocore==1.42.67
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.2
django-storages==1.14.6
idna==3.11
jmespath==1.1.0
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
stripe==14.4.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.6.0
geopy==2.4.0
numpy
redis==7.4.0
django-celery-beat==2.9.0
reportlab==4.2.2
